backend/app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
python-dotenv